+++ src/app/api/auth/route.ts
@@ -17,7 +17,7 @@ export async function GET(): Promise<Response> {
     //set the scope and state for the request
     const scope = 'streaming user-read-private user-read-email user-read-playback-state user-modify-playback-state user-library-read user-read-recently-played user-top-read';
     const state = generateRandomString(16);
-    const authUrl = 'https://accounts.spotify.com/authorize?$'
+    const authUrl = 'https://accounts.spotify.com/authorize?'
     //set the state in a cookie for later verification
     cookies().set('state', state, { httpOnly: true });
     //create the URL for the request
