+++ src/interfaces/artistInterface.ts
