+++ src/interfaces/cardInterface.ts
@@ -1,0 +1,6 @@
+interface CardData {
+    dataRange: string;
+    genreData: Map<string, number>;
+}
+
+export type { CardData }
