+++ src/interfaces/songInterface.ts
